requirements.txt
# Core Dependencies for Streamlit Cloud
groq>=0.4.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# RAG & Embeddings - Optimized for Cloud
sentence-transformers>=2.2.0,<3.0.0
chromadb>=0.4.0,<1.0.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0

# Document Processing
pdfplumber>=0.9.0,<1.0.0
tiktoken>=0.5.0,<1.0.0

# Web UI - Streamlit Cloud Compatible
streamlit>=1.28.0,<2.0.0

# Essential Utilities
numpy>=1.24.0,<2.0.0
requests>=2.28.0,<3.0.0
typing-extensions>=4.0.0,<5.0.0

# Cloud Performance Optimizations
httpx>=0.24.0,<1.0.0
pydantic>=1.10.0,<3.0.0

# Document Processing Extensions (Optional)
# Uncomment if needed:
# pytesseract>=0.3.10,<1.0.0        # For OCR
# python-docx>=0.8.11,<1.0.0        # For DOCX support
# pillow>=10.0.0,<11.0.0            # For image processing